+++ src/components/common/shared/ui/CommonImage/index.tsx
